tune 21

+++ runners/BaseRunner.py
@@ -399,7 +399,7 @@ class BaseRunner(ABC):
                 ### generate data from GP and create dataloader
                 start_time = time.time()
                 if self.config.task.name == 'TFBind8-Exact-v0' or self.config.task.name == 'TFBind10-Exact-v0': 
-                    selected_fit_samples = torch.randperm(self.offline_x.shape[0])[:1/self.config.GP.num_fit_samples_frac*self.offline_x.shape[0]]
+                    selected_fit_samples = torch.randperm(self.offline_x.shape[0])[:int(1/self.config.GP.num_fit_samples_frac*self.offline_x.shape[0])]
                     GP_Model = GP(device=self.config.training.device[0],
                                 x_train=self.offline_x[selected_fit_samples],
                                 y_train=self.offline_y[selected_fit_samples],
